requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Mail==0.9.1
Flask-CORS==3.0.10
Flask-JWT-Extended==4.3.1
Werkzeug==2.0.1
icalendar==4.0.7
python-dotenv==0.19.0
SQLAlchemy==1.4.23
alembic==1.7.1
email-validator==1.1.3